Remove copyright attribution line from setup.py
The file is super-trivial, and thus likely uncopyrightable in the
first place. Additionally, I disagree with the practice of ascribing
copyright to a non-defined entity that is currently in operation
here. The best way to deal with that, other than large amount of
arguing, is to just remove the attribution and the copyright
notice from the file altogether.