Add option to manual configure the worker balance
This commit adds a new cli option to testr run, --worker-file, which
is the path to a yaml file that describes the breakdown of tests per
workers to explicitly instruct testr how to run tests. The yaml file
format is a list of dicts with a single element with a list describing
the contents of a worker as a regex matching test_ids. For example:

  - worker:
    - regex1

  - worker:
    - regex2
    - regex3

would spin up to workers one running tests matching regex1 and the
other worker running tests matching regex2 and regex3.